# metal : optimize dequant q6_K kernel