In this PR I have optimized the Metal q6_K dequantization kernel. The basic idea is to load the weights into 32 bit variables and to do the bit twiddling on them. That gives a total 4-5% speed improvement on prompt processing on my M1 Max.

| Model          | Test   |   t/s master |   t/s dequant_q6_metal |   Speedup |
|:---------------|:-------|-------------:|-----------------------:|----------:|
| llama 8B Q6_K  | pp64   |       289.95 |                 303.06 |      1.05 |
| llama 8B Q6_K  | pp128  |       320.58 |                 335.25 |      1.05 |
| llama 8B Q6_K  | pp256  |       332.92 |                 347.99 |      1.05 |
| llama 8B Q6_K  | pp512  |       339.55 |                 354.55 |      1.04 |
| llama 8B Q6_K  | pp1024 |       337.06 |                 351.88 |      1.04 |
| llama 8B Q6_K  | pp2048 |       330.56 |                 344.86 |      1.04 |
| qwen2 14B Q6_K | pp64   |       153.16 |                 159.54 |      1.04 |
| qwen2 14B Q6_K | pp128  |       169.42 |                 176.74 |      1.04 |
| qwen2 14B Q6_K | pp256  |       177.42 |                 185.07 |      1.04 |
| qwen2 14B Q6_K | pp512  |       179.47 |                 187.02 |      1.04 |
| qwen2 14B Q6_K | pp1024 |       178.11 |                 185.53 |      1.04 |
| qwen2 14B Q6_K | pp2048 |       174.79 |                 181.95 |      1.04 |